# hw2/openapi/marketplace.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  description: API сервиса маркетплейса — управление товарами и заказами
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /auth/register:
    post:
      summary: Регистрация пользователя
      operationId: registerUser
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: Пользователь зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Email уже занят или ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /auth/login:
    post:
      summary: Аутентификация пользователя
      operationId: loginUser
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Неверные учётные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /auth/refresh:
    post:
      summary: Обновление access token
      operationId: refreshToken
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Новый access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Невалидный refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /orders:
    post:
      summary: Создание заказа (USER и ADMIN — да, SELLER — запрещено)
      operationId: createOrder
      tags: [Orders]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Заказ создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Конфликт (активный заказ, неактивный товар, недостаточно остатков)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Ошибка промокода
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '429':
          description: Превышен лимит частоты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /orders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: Получение заказа по ID (USER — только свои, ADMIN — любые, SELLER — запрещено)
      operationId: getOrder
      tags: [Orders]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Заказ найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '403':
          description: Заказ принадлежит другому пользователю
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    put:
      summary: Обновление заказа (USER — только свои, ADMIN — любые, SELLER — запрещено)
      operationId: updateOrder
      tags: [Orders]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdate'
      responses:
        '200':
          description: Заказ обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Заказ принадлежит другому пользователю
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Недопустимое состояние или конфликт
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '429':
          description: Превышен лимит частоты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /orders/{id}/cancel:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    post:
      summary: Отмена заказа (USER — только свои, ADMIN — любые, SELLER — запрещено)
      operationId: cancelOrder
      tags: [Orders]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Заказ отменён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '403':
          description: Заказ принадлежит другому пользователю
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Недопустимое состояние для отмены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token. Роль указана в claim `role`.
  schemas:
    ApiError:
      type: object
      required:
        - error_code
        - message
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
    UserRole:
      type: string
      enum:
        - USER
        - SELLER
        - ADMIN
    UserRegister:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 8
          maxLength: 100
        role:
          $ref: '#/components/schemas/UserRole'
          default: USER
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    UserResponse:
      type: object
      required:
        - id
        - email
        - role
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
    TokenResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - token_type
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
...
    OrderStatus:
      type: string
      enum:
        - CREATED
        - PAYMENT_PENDING
        - PAID
        - SHIPPED
        - COMPLETED
        - CANCELED
    OrderItemInput:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
          maximum: 999
    OrderCreate:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItemInput'
        promo_code:
          type: string
          pattern: '^[A-Z0-9_]{4,20}$'
          nullable: true
    OrderUpdate:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItemInput'
    OrderItemResponse:
      type: object
      required:
        - id
        - product_id
        - quantity
        - price_at_order
      properties:
        id:
          type: integer
          format: int64
        product_id:
          type: integer
          format: int64
        quantity:
          type: integer
        price_at_order:
          type: number
          format: double
    OrderResponse:
      type: object
      required:
        - id
        - user_id
        - status
        - total_amount
        - discount_amount
        - items
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/OrderStatus'
        promo_code_id:
          type: integer
          format: int64
          nullable: true
        total_amount:
          type: number
          format: double
        discount_amount:
          type: number
          format: double
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time